Основной/pip-requirements.txt
python-telegram-bot==12.1.0
pillow==6.0.0
requests==2.21.0
xmltodict==0.12.0
